# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Omastx Machine API
  version: "1.0"
  description: |
    Stable contract for scripts, CI, and AI agents. Authenticate with
    `Authorization: Bearer omx_…` using a token created in Settings.

    Scopes:
    - `read` — fleet, clusters, artifacts, scans, export, settings TTLs
    - `scan` — start a cluster scan

    Token management endpoints require a browser session (cookie + CSRF).
    Admin routes (cluster create/delete, registry credentials, kubeconfigs)
    are session-only and are not part of this machine surface.
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /settings/api-tokens:
    get:
      tags: [Tokens]
      summary: List API tokens
      description: Session + CSRF required. Not available with Bearer tokens.
      security: []
      responses:
        "200":
          description: Active tokens (prefixes only)
          content:
            application/json:
              schema:
                type: object
                required: [tokens]
                properties:
                  tokens:
                    type: array
                    items:
                      $ref: "#/components/schemas/APIToken"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
    post:
      tags: [Tokens]
      summary: Create API token
      description: |
        Session + CSRF required. Returns the plaintext token once in `token`.
        Store it immediately; it cannot be retrieved again.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              properties:
                name:
                  type: string
                  maxLength: 128
                scopes:
                  type: array
                  minItems: 1
                  items:
                    type: string
                    enum: [read, scan]
                expires_in_days:
                  type: integer
                  minimum: 1
                  maximum: 3650
      responses:
        "201":
          description: Token created
          content:
            application/json:
              schema:
                type: object
                required: [token, token_meta]
                properties:
                  token:
                    type: string
                    description: Plaintext Bearer secret (shown once)
                  token_meta:
                    $ref: "#/components/schemas/APIToken"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
  /settings/api-tokens/{id}:
    delete:
      tags: [Tokens]
      summary: Revoke API token
      description: Session + CSRF required.
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Revoked
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: omx_
      description: API token from Settings (`omx_…`)
...
  responses:
    Problem:
      description: RFC 7807-style problem JSON
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      required: [code, title, detail]
      properties:
        code: { type: string }
        title: { type: string }
        detail: { type: string }
...
    APIToken:
      type: object
      required: [id, name, prefix, scopes, created_by, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        prefix: { type: string }
        scopes:
          type: array
          items:
            type: string
            enum: [read, scan]
        created_by: { type: string }
        created_at: { type: string, format: date-time }
        last_used_at:
          type: [string, "null"]
          format: date-time
        expires_at:
          type: [string, "null"]
          format: date-time
...